Changed definition of time_t and clock_t to c_long.  This refs #720.

+++ tango/stdc/time.d
@@ -46,8 +46,8 @@ else
     }
 }
 
-alias int time_t;
-alias int clock_t;
+alias c_long time_t;
+alias c_long clock_t;
 
 version( Win32 )
 {
